feat: Push Notification code refactor and enjancement
